Move functions into domain_model modules to keep the views skinny
